requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-community>=0.3.0
groq>=0.11.0
pydantic>=2.0.0
pdfplumber>=0.11.0
pymupdf>=1.24.0
Pillow>=10.0.0
python-docx>=1.1.0
tavily-python>=0.5.0
python-dotenv>=1.0.0
markitdown>=0.1.0
json-repair>=0.30.0
pyyaml>=6.0.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
